fix(test): fixes tests lacking charset
Signed-off-by: Felipe Zipitria <felipe.zipitria@owasp.org>

While trying to upgrade ftw to py3, we found that some tests were not working.
After some testing, we found that the test has `Content-Type: us-ascii`, but ftw
is not using it unless you specify the `charset`.

The behaviour under py2 was different from py3 when running the tests.
This PR should fix both tests with the expected format for ftw.