Fix false negative in rule 920480
The problem lies in the inner rule. The operator was
"!@rx ^%{tx.allowed_request_content_type_charset}$", which in the
default configuration was expanded to
"!@rx ^utf-8|iso-8859-1|iso-8859-15|windows-1252$".

This pattern is anchored (^..$) but the anchors only apply to the
first and last alternatives, "utf-8" and "windows-1252", respectively,
and not the entire pattern as it lacks the surrounding parenthesis.
This means that "utf-88", "iso-8859-10" or "foo-windows-1252" are
considered valid.

Unfortunately is not possible to just add the parenthesis as
ModSecurity will escape them when resolving the macro.
This solution follows the approach done for 920420, changing the
macro's content format and using @within, making this a
backwards-incompatible change.

Problem noticed by @katef.