Amend test 920100-6: it's no longer an FP Test 920100-6 was previously flagged as being a false positive. With the update to rule 920100 (see commit 5ab8cf58) this is no longer the case.
Test 920100-6 submits a valid HTTP CONNECT request using a domain on the
request line. This was not previously accounted for in rule 920100 and
historically this type of request (and test 920100-6) caused a false
positive. This is no longer the case: such requests are now correctly
handled by CRS, and test 920100-6 works as it should.