fix (regexp-assemble): process all lines for each template
The template processor applies to every following line and can be used
in definitions of other processors. Contrary to the other processors,
the template processor must, therefore, run on all lines and has to run
before any other processor.

Originally, there used to be different types
of processors with file, block, and line scopes; the template processor
broke when this idea was removed in favor of always using block scope.