fix: exclude well known user agents from unix commands
User agents like `curl` and `wget` currently lead to false positives
because they are valid in `User-Agent` headers.
This commit moves the matching against the `User-Agent` header of 932236
into a separate rule 932238, so that words like `curl` and `wget` are
still treated as attacks in other payloads but are excluded for the
`User-Agent` header checks.

In addition, the well known user agents are also excluded from rule
932237, which only matches against `Referer` and `User-Agent`. An
additional rule was omitted for simplicities sake.