fix: use the correct entrypoint defenitions for containers
The docker-entrypoint.sh script for nginx appears to have
changed. It requires arguments that it will `exec` and the first
of these arguments must be `nginx` or `nginx-debug`, otherwise the
templating logic will not be run.

The httpd container now includes a separate executable called
`httpd-foreground`, which is a more convenient than using
`apachectl -D FOREGROUND`.