feat: add dependency checks to PRs and issues (#3298)
GitHub does not (yet) have the ability to block PRs / issues as
dependencies of other PRs issues. This commit uses
gregsdennis/dependency-action as a simple workaround.

Now, wording like "blocks #123" or "depends on #123" in
the PR / issue description will add a check
for the completion of the dependent PR / issue.