feat: replace *-with-params.ra files with suffix replacements
The *-with-params.ra files were conceived as duplicates of the
*-for-cmdline.ra files, to be used in contexts without the `cmdline`
processor. It has been very difficult to keep these files in sync. We
now have a new feature in the crs-toolchain that allows us to get rid of
the duplicate files, the suffix replacements feature of the `include`
and `include-except` processors.

This commit contains one additional change, in order to guarantee
stability. In unix-shell-upto3.ra, some words have been excluded from
being suffixed because they are detected as English words. Without this
change, multiple tests would have to be adjusted and detection would
change significantly for these words (this was an oversight during
original development).