fix: match non-word-boundary of commands with options
Unix commands such as `dig` require options or they won't run. `dig`
uses the `@` anti-evasion pattern (e.g., for 932237). For a command such
as `dig --help`, the final word boundary (`\b`) in 932237 actually
prevents a match. The `@` causes `dig ` to be matched, but the word
boundary will not match because both the space character and the dash
character are in `\W` (for `\b` to match, one of them would have to be
in `\w`).

This commit modifies the final word boundary of 932237 to `(?:\b|\W)`, in
order to fix this.
Note that the final expression must work for commands without options,
commands with options (`@`) and for command prefixes (`~`, e.g., `gcc`
or `pip`).

Fixes #3401