fix: JSON Unicode reflection for tests
httpbin does not properly reflect Unicode sequences but returns JSON
Unicode escape sequences instead. This can break response tests that
rely on the data in the response to be an exact copy of the request
data. See https://github.com/coreruleset/go-ftw/issues/198.

This PR switches the backend for tests from httpbin to the
go-httpbin port, as the port behaves correctly.