fix: separate rule for RCE detection in User-Agent header
Rule 932200 detects RCEs in the User-Agent header (apart from ARGS
etc.). Because the User-Agent header is essentially a free-form field,
932200 produces many FPs with common user agents.

Move the RCE detection in the User-Agent header to a separate rule, akin
to the special detection already implemented for the Referer header.
Only target URL-like values in the User-Agent, as any other detection is
too generic (e.g., the equivalent of rule 932206 doesn't work for the
User-Agent header).

Fixes #3163