fix: prevent FP against brackets in User-Agent header
The Facebook for Android user agent uses a bracket sequence in the user
agent string which triggers FPs for rule 932131. The analysis revealed
that this issue had already been fixed in 932130, the sibling rule with
the same regular expression, but that the regex-assembly file for 932131
did not exist, which is why the regular expression was not updated.

Fixes #3483