fix: remove 'time' prefix from Windows RCE detection
`time` is a Unix command that can be used for RCE evasion, which is why
it is included in the RCE prefix for Unix commands. However, no such
command (or default alias) exists on Windows. In addition, having `time`
in the prefix causes many FPs in natural language, such as in "time for
tea".

fixes #3526