fix: remove .env from lfi-os-files.data (#4024)
* fix: remove .env from lfi-os-files.data

`.env` is probably the most generic entry in `lfi-os-files.data`.
Unfortunatlye, since the words are matched using `@pmFromFile`, `.env`
is easily matched as a substring. Most other entries are fairly unique
or have a prefix or suffix that makes it unlikely for them to become
FPs.

Note that `.env` is only commented out on purpose. `lfi-os-files.data`
is also used as the base for other word files (e.g.,
`restricted-files.data` and since `.env` hasn't been an issue there
until now, we don't want to remove it from those lists.

Fixes #3775

* tests: add FP tests for `.env`