chore: temporarily deactivate test for 920390 (#4027)
We currently override the test result for 920390 to expect an error
because the httpd image segfaults on that test. The new images we are
building no longer segfault, making the test fail. To resolve the
situation, we'll disable the test until the new images have been
released (the images are verified using the CRS test setup, which
currently fails because the test _does not fail_).